+++ myapp/src/ItemDetailBook.tsx
@@ -82,16 +82,22 @@ export const ItemDetailBook = () => {
   };
 
   const DeleteItem = async () => {
-    try {
-      const result = await fetch(
-        `https://uttc-hackathon-backend-4a3g6srehq-uc.a.run.app/items/${item_id}`,
-        { method: "DELETE" }
-      );
-      if (!result.ok) {
-        throw Error(`Failed to delete item: ${result.status}`);
+    const isConfirmed = window.confirm("本当に削除しますか？");
+
+    if (isConfirmed) {
+      try {
+        const result = await fetch(
+          `https://uttc-hackathon-backend-4a3g6srehq-uc.a.run.app/items/${item_id}`,
+          { method: "DELETE" }
+        );
+        if (!result.ok) {
+          throw Error(`Failed to delete item: ${result.status}`);
+        }
+      } catch (err) {
+        console.error(err);
       }
-    } catch (err) {
-      console.error(err);
+    } else {
+      return;
     }
   };
 
@@ -223,7 +229,9 @@ export const ItemDetailBook = () => {
               </div>
 
               <div>
-                <button onClick={DeleteItem}>このアイテムを削除</button>
+                <button onClick={DeleteItem} className="text-red-600">
+                  このアイテムを削除
+                </button>
               </div>
             </div>
             {/* </div> */}
